+++ razorpay/constants/url.py
@@ -1,5 +1,7 @@
 class URL(object):
-    BASE_URL = 'https://api.razorpay.com/v1'
+    BASE_URL = 'https://api.razorpay.com'
+    V1 = '/v1'
+    V2 = '/v2'
     ORDER_URL = "/orders"
     INVOICE_URL = "/invoices"
     PAYMENT_LINK_URL = "/payment_links"
@@ -17,3 +19,13 @@ class URL(object):
     QRCODE_URL = "/payments/qr_codes"
     REGISTRATION_LINK_URL = "/subscription_registration"
     FUND_ACCOUNT_URL = "/fund_accounts"
+    ACCOUNT = "/accounts"
+    STAKEHOLDER = "/stakeholders"
+    PRODUCT = "/products"
+    TNC = "/tnc"
+    TOKEN = "/tokens"
+    IIN = "/iins"
+    WEBHOOK = "/webhooks"
+    DOCUMENT= "/documents"
+    DISPUTE= "/disputes"
+