+++ razorpay/resources/addon.py
@@ -5,10 +5,10 @@ from ..constants.url import URL
 class Addon(Resource):
     def __init__(self, client=None):
         super(Addon, self).__init__(client)
-        self.base_url = URL.ADDON_URL
+        self.base_url = URL.V1 + URL.ADDON_URL
 
     def fetch(self, addon_id, data={}, **kwargs):
-        """"
+        """
         Fetch addon for given Id
 
         Args:
@@ -26,10 +26,12 @@ class Addon(Resource):
         Args:
             addon_id : Id for which addon object has to be deleted
         """
-        return super(Addon, self).delete(addon_id, data, **kwargs)
+        url = '{}/{}'.format(self.base_url, addon_id)
+
+        return self.delete_url(url, data, **kwargs)
 
     def all(self, data={}, **kwargs):
-        """"
+        """
         Fetch all Add-ons
         Returns:
             Dictionary of Add-ons